+++ bidon/src/main/java/org/bidon/sdk/ads/banner/BannerView.kt
@@ -22,6 +22,8 @@ import org.bidon.sdk.adapter.AdSource
 import org.bidon.sdk.adapter.AdViewHolder
 import org.bidon.sdk.adapter.DemandAd
 import org.bidon.sdk.adapter.ext.ad
+import org.bidon.sdk.adapter.ext.notifyExternalLoss
+import org.bidon.sdk.adapter.ext.notifyExternalWin
 import org.bidon.sdk.ads.AdType
 import org.bidon.sdk.ads.AuctionInfo
 import org.bidon.sdk.ads.InitAwaiter
@@ -231,10 +233,7 @@ class BannerView @JvmOverloads constructor(
 
             AdLifecycle.Loaded -> {
                 if (!wasNotified.getAndSet(true)) {
-                    winner?.adSource?.sendLoss(
-                        winnerDemandId = winnerDemandId,
-                        winnerPrice = winnerPrice,
-                    )
+                    winner?.adSource?.notifyExternalLoss(winnerDemandId, winnerPrice)
                     destroyAd()
                 }
             }
@@ -251,8 +250,11 @@ class BannerView @JvmOverloads constructor(
             logInfo(TAG, "Sdk is not initialized")
             return
         }
-        if (adLifecycleFlow.value == AdLifecycle.Loaded && !wasNotified.getAndSet(true)) {
-            winner?.adSource?.sendWin()
+        val adSource = winner?.adSource
+        if (adSource?.canSendWinLoseNotifications() == true &&
+            !wasNotified.getAndSet(true)
+        ) {
+            adSource.notifyExternalWin()
         }
     }