+++ bidon/src/main/java/org/bidon/sdk/ads/interstitial/InterstitialImpl.kt
@@ -14,6 +14,8 @@ import org.bidon.sdk.adapter.AdEvent
 import org.bidon.sdk.adapter.AdSource
 import org.bidon.sdk.adapter.DemandAd
 import org.bidon.sdk.adapter.ext.ad
+import org.bidon.sdk.adapter.ext.notifyExternalLoss
+import org.bidon.sdk.adapter.ext.notifyExternalWin
 import org.bidon.sdk.ads.Ad
 import org.bidon.sdk.ads.AdType
 import org.bidon.sdk.ads.AuctionInfo
@@ -38,6 +40,7 @@ internal class InterstitialImpl(
     Interstitial,
     Extras by demandAd {
     private var userListener: InterstitialListener? = null
+    private var winner: AdSource.Interstitial<*>? = null
     private var observeCallbacksJob: Job? = null
 
     private val adCache: AdCache by lazy {
@@ -114,6 +117,7 @@ internal class InterstitialImpl(
                 logInfo(TAG, "Show failed. No Auction results.")
                 listener.onAdShowFailed(BidonError.AdNotReady)
             } else {
+                winner = adSource
                 adSource.show(activity)
             }
         }
@@ -129,7 +133,7 @@ internal class InterstitialImpl(
             logInfo(TAG, "Sdk is not initialized")
             return
         }
-        adCache.pop()?.adSource?.sendLoss(winnerDemandId, winnerPrice)
+        adCache.pop()?.adSource?.notifyExternalLoss(winnerDemandId, winnerPrice)
         destroyAd()
     }
 
@@ -138,7 +142,9 @@ internal class InterstitialImpl(
             logInfo(TAG, "Sdk is not initialized")
             return
         }
-        adCache.peek()?.adSource?.sendWin()
+        val notifiedSource = winner ?: adCache.peek()?.adSource as? AdSource.Interstitial
+        notifiedSource?.notifyExternalWin()
+        winner = null
     }
 
     override fun destroyAd() {
@@ -150,6 +156,7 @@ internal class InterstitialImpl(
             adCache.clear()
             observeCallbacksJob?.cancel()
             observeCallbacksJob = null
+            winner = null
         }
     }