+++ build-logic/convention-plugins/src/main/kotlin/ComposePlugin.kt
@@ -1,0 +1,17 @@
+import com.android.build.api.dsl.ApplicationExtension
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.gradle.kotlin.dsl.configure
+
+class ComposePlugin : Plugin<Project> {
+    override fun apply(project: Project) = with(project) {
+        pluginManager.apply {
+            apply("org.jetbrains.kotlin.plugin.compose")
+        }
+        extensions.configure<ApplicationExtension> {
+            buildFeatures {
+                compose = true
+            }
+        }
+    }
+}