+++ thirdPartyMediationAdapters/level_play/src/main/java/com/ironsource/adapters/custom/bidon/keeper/AdInstance.kt
@@ -7,6 +7,7 @@ internal interface AdInstance {
     val demandId: String
     val isReady: Boolean
     fun applyAdInfo(ad: Ad): AdInstance
+    fun notifyWin()
     fun notifyLoss(winnerDemandId: String, winnerPrice: Double)
     fun destroy()
 }