+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# 0.10.0 (2025.08.27)
+## Features:
+-[APDM-1462](https://appodeal.atlassian.net/browse/APDM-1462) Promoted 0.10.0 release to stable
+
+# 0.10.0-alpha.1 (2025.08.14)
+## Features:
+- [BDN-1020](https://appodeal.atlassian.net/browse/BDN-1020) Improved Win/Loss notification flow to respect external_win_notifications parameter
+- [BDN-990](https://appodeal.atlassian.net/browse/BDN-990) Updated compileSdkVersion to 35
+- [BDN-954](https://appodeal.atlassian.net/browse/BDN-954) Migrated to build-logic
+- [BDN-1005](https://appodeal.atlassian.net/browse/BDN-1005) Updated Admob adapter to 24.5.0
+- [BDN-1005](https://appodeal.atlassian.net/browse/BDN-1005) Updated GAM adapter to 24.5.0
+- [BDN-1005](https://appodeal.atlassian.net/browse/BDN-1005) Updated Amazon adapter to 11.0.1
+- [BDN-1005](https://appodeal.atlassian.net/browse/BDN-1005) Updated Applovin adapter to 13.3.1
+- [BDN-1005](https://appodeal.atlassian.net/browse/BDN-1005) Updated Bidmachine adapter to 3.4.0
+- [BDN-1005](https://appodeal.atlassian.net/browse/BDN-1005) Updated BigoAds adapter to 5.4.0
+- [BDN-1005](https://appodeal.atlassian.net/browse/BDN-1005) Updated Chartboost adapter to 9.9.1
+- [BDN-1005](https://appodeal.atlassian.net/browse/BDN-1005) Updated DTExchange adapter to 8.3.8
+- [BDN-1005](https://appodeal.atlassian.net/browse/BDN-1005) Updated IronSource adapter to 8.10.0
+- [BDN-1005](https://appodeal.atlassian.net/browse/BDN-1005) Updated Inmobi adapter to 10.8.7
+- [BDN-1005](https://appodeal.atlassian.net/browse/BDN-1005) Updated Mintegral adapter to 16.9.91
+- [BDN-1005](https://appodeal.atlassian.net/browse/BDN-1005) Updated UnityAds adapter to 4.16.0
+- [BDN-1005](https://appodeal.atlassian.net/browse/BDN-1005) Updated Yandex adapter to 7.15.0
+- [BDN-944](https://appodeal.atlassian.net/browse/BDN-944) Remove mediation config from BidMachine configuration
+
 # 0.9.0 (2025.07.24)
 ## Features:
 - [BDN-995](https://appodeal.atlassian.net/browse/BDN-995) Updated Bidmachine SDK to 3.3.0. Added support for BidMachine Placement ID